+++ packages/gluestick/src/config/defaults/glueStickConfig.js
@@ -1,1 +1,7 @@
-module.exports = exports = {};
+module.exports = exports = {
+  host: '0.0.0.0',
+  ports: {
+    client: 8888,
+    server: 8880,
+  },
+};
