--- packages/gluestick/src/config/universalWebpackSettings.json
@@ -1,6 +1,0 @@
-{
-  "server": {
-    "input": "./rendering-service/main.js",
-    "output": "./build/server/server.js"
-  }
-}
