--- packages/gluestick/src/config/webpack.config.client.js
@@ -1,5 +1,0 @@
-const { clientConfiguration } = require('universal-webpack');
-const settings = require('./universalWebpackSettings');
-const configuration = require('./webpack.config');
-
-module.exports = options => clientConfiguration(configuration, settings, options);
