+++ packages/gluestick/src/config/webpack/webpack.config.client.js
@@ -1,0 +1,5 @@
+const { clientConfiguration } = require('universal-webpack');
+
+module.exports = (configuration, settings) => {
+  return options => clientConfiguration(configuration, settings, options);
+};
