+++ packages/gluestick/src/config/webpack/webpack.config.server.js
@@ -1,0 +1,5 @@
+const { serverConfiguration } = require('universal-webpack');
+
+module.exports = (configuration, settings) => {
+  return serverConfiguration(configuration, settings);
+};
